requirements.txt
Flask==2.1.0
google-api-python-client==2.15.0
gunicorn==20.1.0

openai~=0.27.8
tiktoken~=0.4.0
langchain~=0.0.242
youtube-transcript-api==0.4.1
langdetect~=1.0.9
langid~=1.1.6
google-cloud-texttospeech==2.3.0
google-cloud-storage==2.10.0